Correlation analysis

Calculates correlation of a frequency table with an outcome list according to given method. The function currently only works for pairwise correlation, i.e., two categories. Note that the function keyness() is actually better to do the same thing, and this function should not be used anymore in this fashion.

correlationAnalysis(text.ft, categories, method = "spearman",
  culling = 0, ...)

Arguments

text.ft
A matrix, containing words in columns and characters (or plays) in rows. This can be the result of the frequencytable() function.

categories
A factor or numeric vector that represents a list of categories.

method
The correlation method, passed on to cor()

culling
An integer. Words that appear in less items are removed. Defaults to 0 which doesn't remove anything.

...
Arguments passed to cor()

Value

The function returns a data.frame with three columns: The word, it's correlation score, and the category it is correlated to. The latter is mainly for an easier use of the results.

Examples

data(rksp.0)
ft <- frequencytable(rksp.0, byCharacter=TRUE)
g <- factor(c("m","m","m","m","f","m","m","m","f","m","m","f","m"))
rksp.0.cor <- correlationAnalysis(ft, g)

# to pre-filter by the total frequency of a word
ft <- frequencytable(rksp.0, byCharacter=TRUE)
ft <- ft[,colSums(ft) > 5]
correlationAnalysis(ft, g)

#>                      cor         word category
#> mensch       -0.84207536       mensch        f
#> entgegen     -0.73652504     entgegen        f
#> je           -0.72925865           je        f
#> wann         -0.72925865         wann        f
#> bitten       -0.65828059       bitten        f
#> einzig       -0.62380153       einzig        f
#> tone         -0.61307927         tone        f
#> tot          -0.61201656          tot        f
#> vater        -0.60232109        vater        f
#> ihm          -0.59713602          ihm        f
#> merken       -0.59543720       merken        f
#> wo           -0.59543720           wo        f
#> itzt         -0.59160798         itzt        f
#> tat          -0.59160798          tat        f
#> nichts       -0.58960352       nichts        f
#> nein         -0.57890971         nein        f
#> rauber       -0.56720117       rauber        f
#> begegnen     -0.56666667     begegnen        f
#> jeder        -0.56155724        jeder        f
#> ernsthaft    -0.56094261    ernsthaft        f
#> kind         -0.56094261         kind        f
#> verachtung   -0.56094261   verachtung        f
#> zusammen     -0.56094261     zusammen        f
#> halten       -0.54815299       halten        f
#> machen       -0.54427339       machen        f
#> vielleicht   -0.53928447   vielleicht        f
#> weniger      -0.53042899      weniger        f
#> verstand     -0.52927751     verstand        f
#> ha           -0.52856974           ha        f
#> luge         -0.52704628         luge        f
#> ah           -0.50562227           ah        f
#> glauben      -0.50557890      glauben        f
#> dir          -0.50475481          dir        f
#> darum        -0.50393473        darum        f
#> wer          -0.50266539          wer        f
#> mann         -0.48105120         mann        f
#> nehmen       -0.48105120       nehmen        f
#> letzter      -0.48074017      letzter        f
#> verlassen    -0.48074017    verlassen        f
#> warum        -0.48034116        warum        f
#> ob           -0.47963426           ob        f
#> morder       -0.47631080       morder        f
#> eben         -0.47408917         eben        f
#> wir          -0.47340947          wir        f
#> horen        -0.46939178        horen        f
#> gleich       -0.45914138       gleich        f
#> unter        -0.45310103        unter        f
#> wenig        -0.45236007        wenig        f
#> herz         -0.45162272         herz        f
#> denn         -0.44468694         denn        f
#> mein         -0.44036650         mein        f
#> allein       -0.43961496       allein        f
#> wissen       -0.43961496       wissen        f
#> anders       -0.43041423       anders        f
#> graf         -0.42788979         graf        f
#> name         -0.42785314         name        f
#> hab          -0.42418504          hab        f
#> brechen      -0.41890583      brechen        f
#> ehrlich      -0.41890583      ehrlich        f
#> leicht       -0.41890583       leicht        f
#> namlich      -0.41890583      namlich        f
#> wen          -0.41890583          wen        f
#> auf          -0.41821871          auf        f
#> zufall       -0.41718547       zufall        f
#> sprechen     -0.41122462     sprechen        f
#> davon        -0.40720551        davon        f
#> immer        -0.40629960        immer        f
#> erster       -0.40446335       erster        f
#> tag          -0.40111231          tag        f
#> bleiben      -0.39849064      bleiben        f
#> haar         -0.39659041         haar        f
#> ganz         -0.39472173         ganz        f
#> nur          -0.39361761          nur        f
#> von          -0.39306901          von        f
#> mich         -0.39252271         mich        f
#> heissen      -0.39153785      heissen        f
#> her          -0.38818278          her        f
#> wahr         -0.38670386         wahr        f
#> aus          -0.37699904          aus        f
#> heute        -0.37699904        heute        f
#> prinz        -0.37214825        prinz        f
#> tochter      -0.37214825      tochter        f
#> alle         -0.36901651         alle        f
#> herr         -0.36901651         herr        f
#> konnen       -0.36850220       konnen        f
#> hatt         -0.36543533         hatt        f
#> seite        -0.36423767        seite        f
#> also         -0.36227411         also        f
#> appiani      -0.36203585      appiani        f
#> dolch        -0.36055513        dolch        f
#> keiner       -0.36055513       keiner        f
#> erwarten     -0.35872952     erwarten        f
#> verzeihen    -0.35872952    verzeihen        f
#> lieber       -0.35758755       lieber        f
#> anderer      -0.35393559      anderer        f
#> aller        -0.35341472        aller        f
#> da           -0.34882803           da        f
#> kein         -0.34882803         kein        f
#> bei          -0.34733836          bei        f
#> aber         -0.34441541         aber        f
#> auch         -0.34393539         auch        f
#> er           -0.34345737           er        f
#> werden       -0.34298133       werden        f
#> haben        -0.34203518        haben        f
#> ein          -0.34156503          ein        f
#> sein         -0.34156503         sein        f
#> mochten      -0.33683014      mochten        f
#> treffen      -0.33683014      treffen        f
#> wild         -0.33683014         wild        f
#> gehoren      -0.33440687      gehoren        f
#> gewalt       -0.33440687       gewalt        f
#> lange        -0.32914029        lange        f
#> hand         -0.32437680         hand        f
#> weiter       -0.32088985       weiter        f
#> sonst        -0.31983602        sonst        f
#> doch         -0.31892470         doch        f
#> ihr          -0.31848267          ihr        f
#> kommen       -0.31848267       kommen        f
#> seele        -0.31176315        seele        f
#> dosalo       -0.31031645       dosalo        f
#> klein        -0.30820111        klein        f
#> mehr         -0.30472470         mehr        f
#> finden       -0.30382181       finden        f
#> durfen       -0.30337336       durfen        f
#> vor          -0.30337336          vor        f
#> gegen        -0.30203985        gegen        f
#> welcher      -0.29942474      welcher        f
#> messe        -0.29921845        messe        f
#> O            -0.29899545            O        f
#> gott         -0.29860214         gott        f
#> schonheit    -0.29860214    schonheit        f
#> mit          -0.29729660          mit        f
#> sollen       -0.29729660       sollen        f
#> fur          -0.29687639          fur        f
#> nach         -0.29604130         nach        f
#> nun          -0.29604130          nun        f
#> uns          -0.29562638          uns        f
#> ihn          -0.29439203          ihn        f
#> wie          -0.29439203          wie        f
#> hier         -0.29398400         hier        f
#> lassen       -0.29398400       lassen        f
#> wollen       -0.29398400       wollen        f
#> als          -0.29357766          als        f
#> dass         -0.29357766         dass        f
#> nicht        -0.29357766        nicht        f
#> wenn         -0.29357766         wenn        f
#> zu           -0.29357766           zu        f
#> -            -0.29317301            -        f
#> es           -0.29317301           es        f
#> in           -0.29317301           in        f
#> sie          -0.29317301          sie        f
#> so           -0.29317301           so        f
#> und          -0.29317301          und        f
#> was          -0.29317301          was        f
#> der          -0.29277002          der        f
#> ich          -0.29277002          ich        f
#> verstehen    -0.29270357    verstehen        f
#> vergessen    -0.28445674    vergessen        f
#> marinelli    -0.28271693    marinelli        f
#> treten       -0.28069179       treten        f
#> oder         -0.27447267         oder        f
#> sterbend     -0.27247463     sterbend        f
#> an           -0.27099085           an        f
#> sagen        -0.27061211        sagen        f
#> ruhig        -0.26696803        ruhig        f
#> dein         -0.26061578         dein        f
#> nie          -0.26061578          nie        f
#> auge         -0.25859704         auge        f
#> eilen        -0.25495098        eilen        f
#> ergreifen    -0.25495098    ergreifen        f
#> fort         -0.25495098         fort        f
#> indes        -0.25495098        indes        f
#> kennen       -0.25393725       kennen        f
#> man          -0.25133270          man        f
#> daruber      -0.25125584      daruber        f
#> himmel       -0.25125584       himmel        f
#> lang         -0.25125584         lang        f
#> mutter       -0.25096712       mutter        f
#> sich         -0.24397502         sich        f
#> wirklich     -0.24267801     wirklich        f
#> ding         -0.24066739         ding        f
#> wagen        -0.23492271        wagen        f
#> werk         -0.23273733         werk        f
#> leben        -0.23131385        leben        f
#> wert         -0.22537447         wert        f
#> stehen       -0.22333852       stehen        f
#> gut          -0.22297245          gut        f
#> augenblick   -0.21499354   augenblick        f
#> einmal       -0.21357443       einmal        f
#> gewiss       -0.20687763       gewiss        f
#> liebe        -0.20561231        liebe        f
#> grafe        -0.20165493        grafe        f
#> gar          -0.20135990          gar        f
#> unschuldig   -0.19847907   unschuldig        f
#> brautigam    -0.19797975    brautigam        f
#> noch         -0.19736086         noch        f
#> folgen       -0.19648425       folgen        f
#> stunde       -0.19648425       stunde        f
#> dieser       -0.19626135       dieser        f
#> dafur        -0.19612798        dafur        f
#> retten       -0.19612798       retten        f
#> bis          -0.18718575          bis        f
#> etwas        -0.18626591        etwas        f
#> genug        -0.18186184        genug        f
#> viel         -0.17855498         viel        f
#> haus         -0.17802118         haus        f
#> verlieren    -0.17722939    verlieren        f
#> um           -0.17220770           um        f
#> fallen       -0.16720344       fallen        f
#> solcher      -0.16539922      solcher        f
#> fahren       -0.16457015       fahren        f
#> herein       -0.16457015       herein        f
#> nennen       -0.16457015       nennen        f
#> sturzen      -0.16457015      sturzen        f
#> erst         -0.15539822         erst        f
#> uber         -0.15515822         uber        f
#> damit        -0.15468155        damit        f
#> du           -0.15420923           du        f
#> emilia       -0.15420923       emilia        f
#> wort         -0.15420923         wort        f
#> geben        -0.15146345        geben        f
#> ihnen        -0.15079962        ihnen        f
#> ja           -0.15079962           ja        f
#> schon        -0.15079962        schon        f
#> wieder       -0.14907120       wieder        f
#> selbst       -0.14885930       selbst        f
#> marchese     -0.14085904     marchese        f
#> wut          -0.14034589          wut        f
#> welt         -0.14009141         welt        f
#> nahe         -0.13958658         nahe        f
#> gern         -0.13933620         gern        f
#> lustig       -0.13333333       lustig        f
#> schritt      -0.13198650      schritt        f
#> freilich     -0.13051262     freilich        f
#> sehen        -0.12422600        sehen        f
#> daran        -0.12018504        daran        f
#> tief         -0.12018504         tief        f
#> wille        -0.12018504        wille        f
#> wunsch       -0.12018504       wunsch        f
#> geschwind    -0.11944086    geschwind        f
#> lachen       -0.11944086       lachen        f
#> brief        -0.11919585        brief        f
#> liegen       -0.10785689       liegen        f
#> morgen       -0.10343882       morgen        f
#> erlauben     -0.09923953     erlauben        f
#> gering       -0.09923953       gering        f
#> hin          -0.09923953          hin        f
#> voraus       -0.09923953       voraus        f
#> mund         -0.08976553         mund        f
#> spat         -0.08976553         spat        f
#> vermuten     -0.08976553     vermuten        f
#> fassen       -0.08144110       fassen        f
#> wohl         -0.07474886         wohl        f
#> dabei        -0.06599325        dabei        f
#> dazu         -0.06599325         dazu        f
#> lieb         -0.06599325         lieb        f
#> wozu         -0.06599325         wozu        f
#> ablaufen     -0.05972043     ablaufen        f
#> fuss         -0.05665577         fuss        f
#> verdienen    -0.05665577    verdienen        f
#> braut        -0.05603657        braut        f
#> reissen      -0.05603657      reissen        f
#> dich         -0.05563486         dich        f
#> zeit         -0.05348164         zeit        f
#> indem        -0.05171941        indem        f
#> zwar         -0.02986021         zwar        f
#> bloss        -0.02806918        bloss        f
#> tugend       -0.02806918       tugend        f
#> zuruck       -0.02647993       zuruck        f
#> mussen       -0.02513327       mussen        f
#> brauchen      0.00000000     brauchen        f
#> dann          0.00000000         dann        f
#> heut          0.00000000         heut        f
#> kammerherr    0.00000000   kammerherr        f
#> kaum          0.00000000         kaum        f
#> kopf          0.00000000         kopf        f
#> kurz          0.00000000         kurz        f
#> melden        0.00000000       melden        f
#> ohne          0.00000000         ohne        f
#> warten        0.00000000       warten        f
#> tun           0.02491629          tun        m
#> weiss         0.02546964        weiss        m
#> stadt         0.02550785        stadt        m
#> bringen       0.02574081      bringen        m
#> ab            0.02606158           ab        m
#> fuhren        0.02832789       fuhren        m
#> geschehen     0.02986021    geschehen        m
#> glucklich     0.02986021    glucklich        m
#> madchen       0.02986021      madchen        m
#> allerdings    0.02992184   allerdings        m
#> danken        0.02992184       danken        m
#> emilien       0.02992184      emilien        m
#> fehlen        0.02992184       fehlen        m
#> suchen        0.02992184       suchen        m
#> denken        0.05148162       denken        m
#> gnade         0.05295986        gnade        m
#> galotti       0.05339361      galotti        m
#> unglucklich   0.05339361  unglucklich        m
#> lieben        0.05583463       lieben        m
#> beide         0.05959792        beide        m
#> recht         0.07562060        recht        m
#> ehe           0.08498366          ehe        m
#> einer         0.08498366        einer        m
#> fruh          0.08498366         fruh        m
#> gute          0.08498366         gute        m
#> schlimm       0.08498366      schlimm        m
#> wem           0.08498366          wem        m
#> freund        0.10626417       freund        m
#> sehr          0.10678721         sehr        m
#> meinen        0.10971343       meinen        m
#> alt           0.11207313          alt        m
#> endlich       0.11207313      endlich        m
#> stolz         0.11207313        stolz        m
#> verbrechen    0.11207313   verbrechen        m
#> befehlen      0.11227671     befehlen        m
#> gedanke       0.11227671      gedanke        m
#> wurdig        0.11227671       wurdig        m
#> gluck         0.13933620        gluck        m
#> darauf        0.13958658       darauf        m
#> wahrlich      0.13958658     wahrlich        m
#> schicken      0.14085904     schicken        m
#> gehen         0.14885930        gehen        m
#> gnadig        0.14885930       gnadig        m
#> unser         0.15079962        unser        m
#> bruneschi     0.15811388    bruneschi        m
#> claudia       0.15811388      claudia        m
#> gross         0.15965649        gross        m
#> leute         0.16457015        leute        m
#> schlagen      0.16457015     schlagen        m
#> orsina        0.18101793       orsina        m
#> bald          0.18129792         bald        m
#> arm           0.18874956          arm        m
#> weit          0.18874956         weit        m
#> sogleich      0.21392657     sogleich        m
#> weil          0.21499354         weil        m
#> battista      0.23273733     battista        m
#> conti         0.23273733        conti        m
#> gemahlin      0.23273733     gemahlin        m
#> kunstler      0.23273733     kunstler        m
#> portrat       0.23273733      portrat        m
#> ausser        0.24432331       ausser        m
#> grafin        0.25701539       grafin        m
#> mogen         0.28667736        mogen        m
#> weg           0.28667736          weg        m
#> angelo        0.29696962       angelo        m
#> dort          0.29696962         dort        m
#> fall          0.29696962         fall        m
#> jener         0.29696962        jener        m
#> kunst         0.29696962        kunst        m
#> maler         0.29696962        maler        m
#> person        0.29696962       person        m
#> sache         0.29696962        sache        m
#> schwerlich    0.29696962   schwerlich        m
#> vortrefflich  0.29696962 vortrefflich        m
#> zartlich      0.29696962     zartlich        m
#> --            0.29771860           --        m
#> erwagen       0.29771860      erwagen        m
#> furchten      0.29771860     furchten        m
#> guastalla     0.29771860    guastalla        m
#> reden         0.29771860        reden        m
#> sabionetta    0.29771860   sabionetta        m
#> wenigstens    0.29771860   wenigstens        m
#> frau          0.35758755         frau        m
#> bild          0.35832257         bild        m
#> bitte         0.35832257        bitte        m
#> ehre          0.35832257         ehre        m
#> fraulein      0.35832257     fraulein        m
#> freundschaft  0.35832257 freundschaft        m
#> geh           0.35832257          geh        m
#> heraus        0.35832257       heraus        m
#> setzen        0.35832257       setzen        m
#> tod           0.35832257          tod        m
#> kehren        0.35906214       kehren        m
#> umstand       0.36055513      umstand        m
#> verdacht      0.36055513     verdacht        m
#> wegen         0.36055513        wegen        m
#> durch         0.42103768        durch        m