Download preprocessed drama data

This function downloads pre-processed dramatic texts via http and stores them locally in your data directory

installData(dataSource = "tg", dataDirectory = getOption("qd.datadir"),
  downloadSource = "ims", removeZipFile = TRUE,
  baseUrl = "https://github.com/quadrama", remoteUrl = paste0(baseUrl,
  "/data_", dataSource, ".git"))

Arguments

dataSource
Currently, only "tg" (textgrid) is supported

dataDirectory
The directory in which the data is to be stored

downloadSource
No longer used.

removeZipFile
No longer used.

baseUrl
The remote repository owner (e.g., https://github.com/quadrama)

remoteUrl
The URL of the remote repository.

Value

NULL