Isolates Character Speech

isolateCharacterSpeech() isolates the speeches of individual characters and optionally saves them in seperate text files.

isolateCharacterSpeech(drama, segment = c("Drama", "Act", "Scene"),
  minTokenCount = 0, countPunctuation = TRUE, writeToFiles = TRUE,
  dir = getOption("qd.datadir"))

Arguments

drama
A text (or multiple texts, as a QDDrama object)

segment
"Drama", "Act", or "Scene". Determines on what segment-level the speech is isolated.

minTokenCount
The minimal token count for a speech to be considered (default = 0)

countPunctuation
Whether to include punctuation in minTokenCount (default = TRUE)

writeToFiles
Whether to write each isolated speech into a new text file (default = TRUE)

dir
The directory into which the files will be written (default = data directory)

Examples

data(rksp.0)
isolateCharacterSpeech(rksp.0, segment="Scene", writeToFiles=FALSE)

#> $der.prinz_1_1
#>      Token.surface
#>   1:        Klagen
#>   2:             ,
#>   3:        nichts
#>   4:           als
#>   5:        Klagen
#>  ---              
#> 336:             .
#> 337:             -
#> 338:         Steht
#> 339:           auf
#> 340:             .
#> 
#> $der.kammerdiener_1_1
#>     Token.surface
#>  1:          Nein
#>  2:             .
#>  3:          Nach
#>  4:           dem
#>  5:      Marchese
#>  6:           ist
#>  7:     geschickt
#>  8:             .
#>  9:           Und
#> 10:          hier
#> 11:             ,
#> 12:           ein
#> 13:         Brief
#> 14:           von
#> 15:           der
#> 16:        Grafin
#> 17:        Orsina
#> 18:             .
#> 19:           Ihr
#> 20:        Laufer
#> 21:        wartet
#> 22:             .
#> 23:           Sie
#> 24:           ist
#> 25:       gestern
#> 26:            in
#> 27:           die
#> 28:         Stadt
#> 29:      gekommen
#> 30:             .
#> 31:           Der
#> 32:         Maler
#> 33:         Conti
#> 34:          will
#> 35:           die
#> 36:         Gnade
#> 37:         haben
#> 38:             -
#> 39:             -
#>     Token.surface
#> 
#> $der.prinz_1_2
#>     Token.surface
#>  1:         Guten
#>  2:        Morgen
#>  3:             ,
#>  4:         Conti
#>  5:             .
#>  6:           Wie
#>  7:         leben
#>  8:           Sie
#>  9:             ?
#> 10:           Was
#> 11:         macht
#> 12:           die
#> 13:         Kunst
#> 14:             ?
#> 15:           Das
#> 16:          muss
#> 17:           sie
#> 18:         nicht
#> 19:             ;
#> 20:           das
#> 21:          soll
#> 22:           sie
#> 23:         nicht
#> 24:             ,
#> 25:             -
#> 26:            in
#> 27:        meinem
#> 28:       kleinen
#> 29:       Gebiete
#> 30:        gewiss
#> 31:         nicht
#> 32:             .
#> 33:             -
#> 34:          Aber
#> 35:           der
#> 36:      Kunstler
#> 37:          muss
#> 38:          auch
#> 39:      arbeiten
#> 40:        wollen
#> 41:             .
#> 42:           Ich
#> 43:         meine
#> 44:         nicht
#> 45:        vieles
#> 46:             ;
#> 47:       sondern
#> 48:          viel
#> 49:             :
#> 50:           ein
#> 51:       weniges
#> 52:             ;
#> 53:          aber
#> 54:           mit
#> 55:        Fleiss
#> 56:             .
#> 57:             -
#> 58:           Sie
#> 59:        kommen
#> 60:          doch
#> 61:         nicht
#> 62:          leer
#> 63:             ,
#> 64:         Conti
#> 65:             ?
#> 66:         Jenes
#> 67:           ist
#> 68:             ?
#> 69:             -
#> 70:          Kann
#> 71:           ich
#> 72:          mich
#> 73:          doch
#> 74:          kaum
#> 75:      erinnern
#> 76:             -
#> 77:          Wahr
#> 78:             !
#> 79:             -
#> 80:           Der
#> 81:       Auftrag
#> 82:           ist
#> 83:           nur
#> 84:           ein
#> 85:         wenig
#> 86:           von
#> 87:         lange
#> 88:           her
#> 89:             .
#> 90:            Wo
#> 91:          sind
#> 92:           die
#> 93:        Stucke
#> 94:             ?
#>     Token.surface
#> 
#> $conti_1_2
#>      Token.surface
#>   1:         Prinz
#>   2:             ,
#>   3:           die
#>   4:         Kunst
#>   5:          geht
#>  ---              
#>  97:             :
#>  98:           ich
#>  99:          hole
#> 100:           sie
#> 101:             .
#> 
#> $der.prinz_1_3
#>      Token.surface
#>   1:           Ihr
#>   2:          Bild
#>   3:             !
#>   4:             -
#>   5:           mag
#>  ---              
#> 144:           ich
#> 145:           bin
#> 146:            so
#> 147:        besser
#> 148:             .
#> 
#> $conti_1_4
#>      Token.surface
#>   1:           Ich
#>   2:         bitte
#>   3:             ,
#>   4:         Prinz
#>   5:             ,
#>  ---              
#> 659:             ,
#> 660:           als
#> 661:           die
#> 662:         Kunst
#> 663:             .
#> 
#> $der.prinz_1_4
#>      Token.surface
#>   1:  Vortrefflich
#>   2:             ,
#>   3:         Conti
#>   4:             ;
#>   5:             -
#>  ---              
#> 811:             .
#> 812:         Conti
#> 813:          geht
#> 814:            ab
#> 815:             .
#> 
#> $der.prinz_1_5
#>      Token.surface
#>   1:            So
#>   2:          viel
#>   3:            er
#>   4:          will
#>   5:             !
#>  ---              
#> 153:         konnt
#> 154:             '
#> 155:           ich
#> 156:         haben
#> 157:             !
#> 
#> $marinelli_1_6
#>       Token.surface
#>    1:      Gnadiger
#>    2:          Herr
#>    3:             ,
#>    4:           Sie
#>    5:        werden
#>   ---              
#> 1058:             .
#> 1059:             -
#> 1060:     Verstehen
#> 1061:           Sie
#> 1062:             ?
#> 
#> $der.prinz_1_6
#>      Token.surface
#>   1:           Ich
#>   2:         bekam
#>   3:          Lust
#>   4:             ,
#>   5:   auszufahren
#>  ---              
#> 946:             .
#> 947:     Marinelli
#> 948:          geht
#> 949:            ab
#> 950:             .
#> 
#> $der.prinz_1_7
#>      Token.surface
#>   1:      Sogleich
#>   2:             !
#>   3:      sogleich
#>   4:             !
#>   5:             -
#>  ---              
#> 294:            wo
#> 295:         deine
#> 296: Vorsprecherin
#> 297:             -
#> 298:             -
#> 
#> $der.kammerdiener_1_7
#>    Token.surface
#> 1:       Camillo
#> 2:          Rota
#> 3:             .
#> 
#> $der.prinz_1_8
#>      Token.surface
#>   1:        Kommen
#>   2:           Sie
#>   3:             ,
#>   4:          Rota
#>   5:             ,
#>  ---              
#> 141:        mehres
#> 142:             !
#> 143:          Geht
#> 144:            ab
#> 145:             .
#> 
#> $camillo.rota_1_8
#>      Token.surface
#>   1:           Gut
#>   2:             ,
#>   3:      gnadiger
#>   4:          Herr
#>   5:             .
#>  ---              
#> 102:        dieses
#> 103:    grassliche
#> 104:         Recht
#> 105:          gern
#> 106:             !
#> 
#> $claudia.galotti_2_1
#>     Token.surface
#>  1:           Wer
#>  2:      sprengte
#>  3:            da
#>  4:            in
#>  5:           den
#>  6:           Hof
#>  7:             ?
#>  8:          Mein
#>  9:        Gemahl
#> 10:             ?
#> 11:           Ist
#> 12:            es
#> 13:       moglich
#> 14:             ?
#> 15:            So
#> 16:    unvermutet
#> 17:             ?
#> 18:             -
#> 19:           Ihm
#> 20:      entgegen
#> 21:        eilend
#> 22:             .
#> 23:            Ah
#> 24:             !
#> 25:          mein
#> 26:        Bester
#> 27:             !
#> 28:             -
#>     Token.surface
#> 
#> $pirro_2_1
#>     Token.surface
#>  1:         Unser
#>  2:          Herr
#>  3:             ,
#>  4:       gnadige
#>  5:          Frau
#>  6:             .
#>  7:            Er
#>  8:         folgt
#>  9:           mir
#> 10:           auf
#> 11:           dem
#> 12:         Fusse
#> 13:             .
#> 
#> $odoardo_2_2
#>      Token.surface
#>   1:         Guten
#>   2:        Morgen
#>   3:             ,
#>   4:         meine
#>   5:         Liebe
#>  ---              
#> 112:        allein
#> 113:      gegangen
#> 114:          sein
#> 115:             .
#> 116:             -
#> 
#> $claudia.galotti_2_2
#>      Token.surface
#>   1:           Und
#>   2:           auf
#>   3:           die
#>   4:   angenehmste
#>   5:           Art
#>  ---              
#> 106:           auf
#> 107:         heute
#> 108:            zu
#> 109:     verbitten
#> 110:             .
#> 
#> $pirro_2_3
#>      Token.surface
#>   1:           Die
#>   2:          sich
#>   3:           nur
#>   4:           aus
#>   5:     Neugierde
#>  ---              
#> 313:          ewig
#> 314:             !
#> 315:           Ich
#> 316: Unglucklicher
#> 317:             !
#> 
#> $angelo_2_3
#>      Token.surface
#>   1:         Pirro
#>   2:             !
#>   3:             -
#>   4:         Pirro
#>   5:             !
#>  ---              
#> 412:        kannst
#> 413:             !
#> 414:          Geht
#> 415:            ab
#> 416:             .
#> 
#> $odoardo_2_4
#>      Token.surface
#>   1:           Sie
#>   2:        bleibt
#>   3:           mir
#>   4:            zu
#>   5:          lang
#>  ---              
#> 502:             -
#> 503:         Kommt
#> 504:     glucklich
#> 505:          nach
#> 506:             !
#> 
#> $claudia.galotti_2_4
#>      Token.surface
#>   1:          Noch
#>   2:         einen
#>   3:    Augenblick
#>   4:             ,
#>   5:       Odoardo
#>  ---              
#> 179:             -
#> 180:             -
#> 181:           Wie
#> 182:            so
#> 183:             ?
#> 
#> $pirro_2_4
#>    Token.surface
#> 1:          Hier
#> 2:           bin
#> 3:           ich
#> 4:             .
#> 
#> $claudia.galotti_2_5
#>     Token.surface
#>  1:         Welch
#>  2:           ein
#>  3:          Mann
#>  4:             !
#>  5:             -
#>  6:             O
#>  7:             ,
#>  8:           der
#>  9:        rauhen
#> 10:        Tugend
#> 11:             !
#> 12:             -
#> 13:          wenn
#> 14:        anders
#> 15:           sie
#> 16:        diesen
#> 17:         Namen
#> 18:     verdienet
#> 19:             .
#> 20:             -
#> 21:         Alles
#> 22:       scheint
#> 23:           ihr
#> 24:    verdachtig
#> 25:             ,
#> 26:         alles
#> 27:      strafbar
#> 28:             !
#> 29:             -
#> 30:          Oder
#> 31:             ,
#> 32:          wenn
#> 33:           das
#> 34:           die
#> 35:      Menschen
#> 36:        kennen
#> 37:        heisst
#> 38:             ;
#> 39:             -
#> 40:           wer
#> 41:        sollte
#> 42:          sich
#> 43:      wunschen
#> 44:             ,
#> 45:           sie
#> 46:            zu
#> 47:        kennen
#> 48:             ?
#> 49:             -
#> 50:            Wo
#> 51:        bleibt
#> 52:          aber
#> 53:          auch
#> 54:        Emilia
#> 55:             ?
#> 56:             -
#> 57:            Er
#> 58:           ist
#> 59:           des
#> 60:        Vaters
#> 61:         Feind
#> 62:             :
#> 63:      folglich
#> 64:             -
#> 65:      folglich
#> 66:             ,
#> 67:          wenn
#> 68:            er
#> 69:           ein
#> 70:          Auge
#> 71:           fur
#> 72:           die
#> 73:       Tochter
#> 74:           hat
#> 75:             ,
#> 76:            so
#> 77:           ist
#> 78:            es
#> 79:        einzig
#> 80:             ,
#> 81:            um
#> 82:           ihn
#> 83:            zu
#> 84:   beschimpfen
#> 85:             ?
#> 86:             -
#>     Token.surface
#> 
#> $emilia_2_6
#>      Token.surface
#>   1:          Wohl
#>   2:           mir
#>   3:             !
#>   4:          wohl
#>   5:           mir
#>  ---              
#> 929:            Es
#> 930:           ist
#> 931:          sein
#> 932:          Gang
#> 933:             .
#> 
#> $claudia.galotti_2_6
#>      Token.surface
#>   1:           Was
#>   2:           ist
#>   3:           dir
#>   4:             ,
#>   5:         meine
#>  ---              
#> 616:            so
#> 617:          viel
#> 618:           als
#> 619:        nichts
#> 620:             .
#> 
#> $appiani_2_7
#>      Token.surface
#>   1:            Ah
#>   2:             ,
#>   3:         meine
#>   4:      Teuerste
#>   5:             !
#>  ---              
#> 374:           sie
#> 375:           die
#> 376:         Natur
#> 377:        schlug
#> 378:             -
#> 
#> $emilia_2_7
#>      Token.surface
#>   1:           Ich
#>   2:      wunschte
#>   3:           Sie
#>   4:        heiter
#>   5:             ,
#>  ---              
#> 252:            so
#> 253:           vor
#> 254:         Ihnen
#> 255:            da
#> 256:             !
#> 
#> $claudia.galotti_2_7
#>      Token.surface
#>   1:           Der
#>   2:          mein
#>   3:      grosster
#>   4:         Stolz
#>   5:          sein
#>  ---              
#> 120:           als
#> 121:           von
#> 122:       Steinen
#> 123:             ?
#> 124:             -
#> 
#> $appiani_2_8
#>      Token.surface
#>   1:        Perlen
#>   2:      bedeuten
#>   3:        Tranen
#>   4:             !
#>   5:             -
#>  ---              
#> 230:          noch
#> 231:           bei
#> 232:           ihm
#> 233:     vorfahren
#> 234:             .
#> 
#> $claudia.galotti_2_8
#>     Token.surface
#>  1:      Emiliens
#>  2:   Beobachtung
#>  3:             ,
#>  4:          Herr
#>  5:          Graf
#>  6:             ,
#>  7:           war
#>  8:            so
#>  9:       schnell
#> 10:             ,
#> 11:           als
#> 12:       richtig
#> 13:             .
#> 14:           Sie
#> 15:          sind
#> 16:          heut
#> 17:             '
#> 18:       ernster
#> 19:           als
#> 20:    gewohnlich
#> 21:             .
#> 22:           Nur
#> 23:          noch
#> 24:         einen
#> 25:       Schritt
#> 26:           von
#> 27:           dem
#> 28:         Ziele
#> 29:         Ihrer
#> 30:       Wunsche
#> 31:             ,
#> 32:             -
#> 33:         sollt
#> 34:             '
#> 35:            es
#> 36:           Sie
#> 37:         reuen
#> 38:             ,
#> 39:          Herr
#> 40:          Graf
#> 41:             ,
#> 42:          dass
#> 43:            es
#> 44:           das
#> 45:          Ziel
#> 46:         Ihrer
#> 47:       Wunsche
#> 48:       gewesen
#> 49:             ?
#> 50:           Sie
#> 51:        machen
#> 52:          mich
#> 53:       unruhig
#> 54:             ,
#> 55:          Herr
#> 56:          Graf
#> 57:             -
#> 58:           Wie
#> 59:            so
#> 60:             ?
#> 61:           Bei
#> 62:           dem
#> 63:       Prinzen
#> 64:             ?
#>     Token.surface
#> 
#> $pirro_2_9
#>     Token.surface
#>  1:       Gnadige
#>  2:          Frau
#>  3:             ,
#>  4:           der
#>  5:      Marchese
#>  6:     Marinelli
#>  7:          halt
#>  8:           vor
#>  9:           dem
#> 10:         Hause
#> 11:             ,
#> 12:           und
#> 13:    erkundiget
#> 14:          sich
#> 15:          nach
#> 16:           dem
#> 17:         Herrn
#> 18:        Grafen
#> 19:             .
#> 20:          Hier
#> 21:           ist
#> 22:            er
#> 23:         schon
#> 24:             .
#> 25:        Offnet
#> 26:           ihm
#> 27:           die
#> 28:          Ture
#> 29:           und
#> 30:          geht
#> 31:            ab
#> 32:             .
#>     Token.surface
#> 
#> $appiani_2_9
#>    Token.surface
#> 1:          Nach
#> 2:           mir
#> 3:             ?
#> 
#> $marinelli_2_9
#>     Token.surface
#>  1:           Ich
#>  2:          bitt
#>  3:             '
#>  4:            um
#>  5:    Verzeihung
#>  6:             ,
#>  7:       gnadige
#>  8:          Frau
#>  9:             .
#> 10:             -
#> 11:          Mein
#> 12:          Herr
#> 13:          Graf
#> 14:             ,
#> 15:           ich
#> 16:           war
#> 17:           vor
#> 18:         Ihrem
#> 19:         Hause
#> 20:             ,
#> 21:           und
#> 22:        erfuhr
#> 23:             ,
#> 24:          dass
#> 25:           ich
#> 26:           Sie
#> 27:          hier
#> 28:       treffen
#> 29:         wurde
#> 30:             .
#> 31:           Ich
#> 32:           hab
#> 33:             '
#> 34:           ein
#> 35:    dringendes
#> 36:      Geschaft
#> 37:            an
#> 38:           Sie
#> 39:             -
#> 40:       Gnadige
#> 41:          Frau
#> 42:             ,
#> 43:           ich
#> 44:         bitte
#> 45:      nochmals
#> 46:            um
#> 47:    Verzeihung
#> 48:             ;
#> 49:            es
#> 50:           ist
#> 51:            in
#> 52:       einigen
#> 53:       Minuten
#> 54:     geschehen
#> 55:             .
#>     Token.surface
#> 
#> $claudia.galotti_2_9
#>     Token.surface
#>  1:           Die
#>  2:           ich
#>  3:         nicht
#>  4:     verzogern
#>  5:          will
#>  6:             .
#>  7:         Macht
#>  8:           ihm
#>  9:          eine
#> 10:    Verbeugung
#> 11:           und
#> 12:          geht
#> 13:            ab
#> 14:             .
#> 
#> $appiani_2_10
#>      Token.surface
#>   1:           Nun
#>   2:             ,
#>   3:          mein
#>   4:          Herr
#>   5:             ?
#>  ---              
#> 419:           Sie
#> 420:             ,
#> 421:        kommen
#> 422:           Sie
#> 423:             !
#> 
#> $marinelli_2_10
#>      Token.surface
#>   1:           Ich
#>   2:         komme
#>   3:           von
#>   4:           des
#>   5:       Prinzen
#>  ---              
#> 468:          Graf
#> 469:             ,
#> 470:           nur
#> 471:        Geduld
#> 472:             !
#> 
#> $appiani_2_11
#>      Token.surface
#>  1:            Geh
#>  2:              ,
#>  3: Nichtswurdiger
#>  4:              !
#>  5:              -
#>  6:             Ha
#>  7:              !
#>  8:            das
#>  9:            hat
#> 10:            gut
#> 11:          getan
#> 12:              .
#> 13:           Mein
#> 14:           Blut
#> 15:            ist
#> 16:             in
#> 17:        Wallung
#> 18:       gekommen
#> 19:              .
#> 20:            Ich
#> 21:          fuhle
#> 22:           mich
#> 23:         anders
#> 24:            und
#> 25:         besser
#> 26:              .
#> 27:         Nichts
#> 28:              ,
#> 29:        gnadige
#> 30:           Frau
#> 31:              ,
#> 32:            gar
#> 33:         nichts
#> 34:              .
#> 35:            Der
#> 36:     Kammerherr
#> 37:      Marinelli
#> 38:            hat
#> 39:            mir
#> 40:          einen
#> 41:        grossen
#> 42:         Dienst
#> 43:       erwiesen
#> 44:              .
#> 45:             Er
#> 46:            hat
#> 47:           mich
#> 48:            des
#> 49:         Ganges
#> 50:            zum
#> 51:        Prinzen
#> 52:      uberhoben
#> 53:              .
#> 54:            Wir
#> 55:         konnen
#> 56:            nun
#> 57:             um
#> 58:             so
#> 59:           viel
#> 60:         fruher
#> 61:       abfahren
#> 62:              .
#> 63:            Ich
#> 64:           gehe
#> 65:              ,
#> 66:          meine
#> 67:          Leute
#> 68:             zu
#> 69:        treiben
#> 70:              ,
#> 71:            und
#> 72:            bin
#> 73:       sogleich
#> 74:         wieder
#> 75:           hier
#> 76:              .
#> 77:         Emilia
#> 78:           wird
#> 79:          indes
#> 80:           auch
#> 81:         fertig
#> 82:              .
#> 83:           Ganz
#> 84:          ruhig
#> 85:              ,
#> 86:        gnadige
#> 87:           Frau
#> 88:              .
#> 89:            Sie
#> 90:           geht
#> 91:         herein
#> 92:            und
#> 93:             er
#> 94:           fort
#> 95:              .
#>      Token.surface
#> 
#> $claudia.galotti_2_11
#>     Token.surface
#>  1:          Gott
#>  2:             !
#>  3:          Herr
#>  4:          Graf
#>  5:             -
#>  6:           Ich
#>  7:           hab
#>  8:             '
#>  9:         einen
#> 10:      heftigen
#> 11:   Wortwechsel
#> 12:        gehort
#> 13:             .
#> 14:             -
#> 15:           Ihr
#> 16:       Gesicht
#> 17:        gluhet
#> 18:             .
#> 19:           Was
#> 20:           ist
#> 21:   vorgefallen
#> 22:             ?
#> 23:            In
#> 24:           der
#> 25:           Tat
#> 26:             ?
#> 27:          Kann
#> 28:           ich
#> 29:          ganz
#> 30:         ruhig
#> 31:          sein
#> 32:             ,
#> 33:          Herr
#> 34:          Graf
#> 35:             ?
#>     Token.surface
#> 
#> $marinelli_3_1
#>      Token.surface
#>   1:       Umsonst
#>   2:             ;
#>   3:            er
#>   4:        schlug
#>   5:           die
#>  ---              
#> 688:           Der
#> 689:         Prinz
#> 690:          geht
#> 691:            ab
#> 692:             .
#> 
#> $der.prinz_3_1
#>      Token.surface
#>   1:           Und
#>   2:            so
#>   3:        bleibt
#>   4:            es
#>   5:         dabei
#>  ---              
#> 375:           dem
#> 376:         nicht
#> 377:            ab
#> 378:             -
#> 379:             -
#> 
#> $marinelli_3_2
#>      Token.surface
#>   1:          Dort
#>   2:         fahrt
#>   3:           der
#>   4:         Wagen
#>   5:       langsam
#>  ---              
#> 310:           ich
#> 311:            um
#> 312:           die
#> 313:    Gewissheit
#> 314:             !
#> 
#> $angelo_3_2
#>      Token.surface
#>   1:        Passen
#>   2:           Sie
#>   3:           auf
#>   4:             ,
#>   5:          Herr
#>  ---              
#> 258:        andere
#> 259:             .
#> 260:          Geht
#> 261:            ab
#> 262:             .
#> 
#> $der.prinz_3_3
#>      Token.surface
#>   1:          Dort
#>   2:         kommt
#>   3:           sie
#>   4:             ,
#>   5:           die
#>  ---              
#> 234:          mich
#> 235:          mehr
#> 236:     gesammelt
#> 237:          habe
#> 238:             .
#> 
#> $marinelli_3_3
#>     Token.surface
#>  1:            So
#>  2:         haben
#>  3:           wir
#>  4:           sie
#>  5:          doch
#>  6:          furs
#>  7:         erste
#>  8:             .
#>  9:           Auf
#> 10:           das
#> 11:         alles
#> 12:         weiss
#> 13:           ich
#> 14:      freilich
#> 15:          noch
#> 16:        nichts
#> 17:            zu
#> 18:     antworten
#> 19:             .
#> 20:          Aber
#> 21:           wir
#> 22:        mussen
#> 23:         sehen
#> 24:             .
#> 25:      Gedulden
#> 26:           Sie
#> 27:          sich
#> 28:             ,
#> 29:      gnadiger
#> 30:          Herr
#> 31:             .
#> 32:           Der
#> 33:         erste
#> 34:       Schritt
#> 35:        musste
#> 36:          doch
#> 37:         getan
#> 38:          sein
#> 39:             .
#> 40:             -
#> 41:    Vielleicht
#> 42:        mussen
#> 43:           wir
#> 44:         nicht
#> 45:             .
#> 46:             -
#> 47:            Da
#> 48:          sind
#> 49:       tausend
#> 50:         Dinge
#> 51:             ,
#> 52:           auf
#> 53:           die
#> 54:          sich
#> 55:        weiter
#> 56:        fussen
#> 57:         lasst
#> 58:             .
#> 59:             -
#> 60:           Und
#> 61:     vergessen
#> 62:           Sie
#> 63:          denn
#> 64:           das
#> 65:    Vornehmste
#> 66:             ?
#> 67:           Die
#> 68:         Kunst
#> 69:            zu
#> 70:      gefallen
#> 71:             ,
#> 72:            zu
#> 73:     uberreden
#> 74:             ,
#> 75:             -
#> 76:           die
#> 77:         einem
#> 78:       Prinzen
#> 79:             ,
#> 80:       welcher
#> 81:         liebt
#> 82:             ,
#> 83:           nie
#> 84:        fehlet
#> 85:             .
#>     Token.surface
#> 
#> $marinelli_3_4
#>      Token.surface
#>   1:          Wenn
#>   2:           sie
#>   3:           ihn
#>   4:         nicht
#>   5:        selbst
#>  ---              
#> 277:          Hier
#> 278:           ist
#> 279:            er
#> 280:         schon
#> 281:             .
#> 
#> $battista_3_4
#>     Token.surface
#>  1:           Nur
#>  2:          hier
#>  3:        herein
#>  4:             ,
#>  5:      gnadiges
#>  6:      Fraulein
#>  7:             .
#>  8:           Ich
#>  9:       vermute
#> 10:             .
#> 11:    Geschossen
#> 12:             ?
#> 13:             -
#> 14:           Das
#> 15:          ware
#> 16:             !
#> 17:             -
#> 18:           Ich
#> 19:          will
#> 20:        gleich
#> 21:          nach
#> 22:         ihnen
#> 23:      ausgehen
#> 24:             .
#>     Token.surface
#> 
#> $emilia_3_4
#>      Token.surface
#>   1:            Ah
#>   2:             !
#>   3:             -
#>   4:            Ah
#>   5:             !
#>  ---              
#> 303:            in
#> 304:  Gesellschaft
#> 305:        meiner
#> 306:        Mutter
#> 307:             ?
#> 
#> $der.prinz_3_5
#>      Token.surface
#>   1:            Wo
#>   2:           ist
#>   3:           sie
#>   4:             ?
#>   5:            wo
#>  ---              
#> 344:           uns
#> 345:             ,
#> 346:     Marinelli
#> 347:             .
#> 348:             -
#> 
#> $emilia_3_5
#>      Token.surface
#>   1:            Ah
#>   2:             ,
#>   3:    gnadigster
#>   4:          Herr
#>   5:             !
#>   6:            wo
#>   7:          sind
#>   8:           sie
#>   9:             ?
#>  10:            Wo
#>  11:           ist
#>  12:         meine
#>  13:        Mutter
#>  14:             ?
#>  15:          Gott
#>  16:             ,
#>  17:            in
#>  18:       welchem
#>  19:      Zustande
#>  20:         werde
#>  21:           ich
#>  22:           die
#>  23:          eine
#>  24:             ,
#>  25:          oder
#>  26:           den
#>  27:        andern
#>  28:             ,
#>  29:    vielleicht
#>  30:       treffen
#>  31:             !
#>  32:          Ganz
#>  33:        gewiss
#>  34:       treffen
#>  35:             !
#>  36:             -
#>  37:          denn
#>  38:           Sie
#>  39:     verhehlen
#>  40:           mir
#>  41:             ,
#>  42:      gnadiger
#>  43:          Herr
#>  44:             -
#>  45:           ich
#>  46:           seh
#>  47:             '
#>  48:            es
#>  49:             ,
#>  50:           Sie
#>  51:     verhehlen
#>  52:           mir
#>  53:             -
#>  54:          Aber
#>  55:             -
#>  56:          wenn
#>  57:         ihnen
#>  58:        nichts
#>  59:   widerfahren
#>  60:             -
#>  61:          wenn
#>  62:         meine
#>  63:      Ahnungen
#>  64:          mich
#>  65:        trugen
#>  66:             ;
#>  67:             -
#>  68:         warum
#>  69:          sind
#>  70:           sie
#>  71:         nicht
#>  72:         schon
#>  73:          hier
#>  74:             ?
#>  75:         Warum
#>  76:         kamen
#>  77:           sie
#>  78:         nicht
#>  79:           mit
#>  80:         Ihnen
#>  81:             ,
#>  82:      gnadiger
#>  83:          Herr
#>  84:             ?
#>  85:           Was
#>  86:          soll
#>  87:           ich
#>  88:           tun
#>  89:             !
#>  90:           Die
#>  91:         Hande
#>  92:       ringend
#>  93:             .
#>  94:            Zu
#>  95:         Ihren
#>  96:        Fussen
#>  97:             ,
#>  98:      gnadiger
#>  99:          Herr
#> 100:             -
#>      Token.surface
#> 
#> $marinelli_3_5
#>      Token.surface
#>   1:        Folgen
#>   2:           Sie
#>   3:           uns
#>   4:             ,
#>   5:             -
#>  ---              
#> 104:      Battista
#> 105:             ?
#> 106:           was
#> 107:         gibts
#> 108:             ?
#> 
#> $battista_3_6
#>      Token.surface
#>   1:           Die
#>   2:        Mutter
#>   3:             ,
#>   4:          Herr
#>   5:    Kammerherr
#>  ---              
#> 122:           Sie
#> 123:             !
#> 124:         horen
#> 125:           Sie
#> 126:             !
#> 
#> $marinelli_3_6
#>      Token.surface
#>   1:         Dacht
#>   2:             '
#>   3:          ichs
#>   4:          doch
#>   5:             !
#>  ---              
#> 159:   neugierigen
#> 160:     Begleiter
#> 161:            zu
#> 162:     entfernen
#> 163:             .
#> 
#> $claudia.galotti_3_6
#>     Token.surface
#>  1:        Emilia
#>  2:             !
#>  3:        Emilia
#>  4:             !
#>  5:          Mein
#>  6:          Kind
#>  7:             ,
#>  8:            wo
#>  9:          bist
#> 10:            du
#> 11:             ?
#> 
#> $claudia.galotti_3_7
#>     Token.surface
#>  1:            Ha
#>  2:             !
#>  3:           der
#>  4:           hob
#>  5:           sie
#>  6:           aus
#>  7:           dem
#>  8:         Wagen
#>  9:             !
#> 10:           Der
#> 11:        fuhrte
#> 12:           sie
#> 13:          fort
#> 14:             !
#> 15:           Ich
#> 16:       erkenne
#> 17:          dich
#> 18:             .
#> 19:            Wo
#> 20:           ist
#> 21:           sie
#> 22:             ?
#> 23:        Sprich
#> 24:             ,
#> 25: Unglucklicher
#> 26:             !
#> 27:             O
#> 28:             ,
#> 29:          wenn
#> 30:            du
#> 31:          Dank
#> 32:    verdienest
#> 33:             :
#> 34:            In
#> 35:         einem
#> 36:      gelinden
#> 37:          Tone
#> 38:             .
#> 39:             -
#> 40:            so
#> 41:      verzeihe
#> 42:           mir
#> 43:             ,
#> 44:     ehrlicher
#> 45:          Mann
#> 46:             !
#> 47:             -
#> 48:            Wo
#> 49:           ist
#> 50:           sie
#> 51:             ?
#> 52:             -
#> 53:         Lasst
#> 54:          mich
#> 55:           sie
#> 56:         nicht
#> 57:        langer
#> 58:     entbehren
#> 59:             .
#> 60:            Wo
#> 61:           ist
#> 62:           sie
#> 63:             ?
#>     Token.surface
#> 
#> $battista_3_7
#>     Token.surface
#>  1:           Das
#>  2:           ist
#>  3:          mein
#>  4:          Dank
#>  5:             ?
#>  6:             O
#>  7:             ,
#>  8:          Ihre
#>  9:        Gnaden
#> 10:             ,
#> 11:           sie
#> 12:        konnte
#> 13:            in
#> 14:           dem
#> 15:       Schosse
#> 16:           der
#> 17:     Seligkeit
#> 18:         nicht
#> 19:   aufgehobner
#> 20:          sein
#> 21:             .
#> 22:             -
#> 23:          Hier
#> 24:          mein
#> 25:          Herr
#> 26:          wird
#> 27:          Ihre
#> 28:        Gnaden
#> 29:            zu
#> 30:           ihr
#> 31:        fuhren
#> 32:             .
#> 33:         Gegen
#> 34:        einige
#> 35:         Leute
#> 36:             ,
#> 37:        welche
#> 38:   nachdringen
#> 39:        wollen
#> 40:             .
#> 41:        Zuruck
#> 42:            da
#> 43:             !
#> 44:           ihr
#> 45:             !
#>     Token.surface
#> 
#> $claudia.galotti_3_8
#>      Token.surface
#>   1:          Dein
#>   2:          Herr
#>   3:             ?
#>   4:             -
#>   5:      Erblickt
#>  ---              
#> 563:           und
#> 564:     Marinelli
#> 565:           ihr
#> 566:          nach
#> 567:             .
#> 
#> $marinelli_3_8
#>      Token.surface
#>   1:           Mit
#>   2:        vielem
#>   3:     Vergnugen
#>   4:             ,
#>   5:       gnadige
#>  ---              
#> 210:             ,
#> 211:            wo
#> 212:           Sie
#> 213:          sind
#> 214:             .
#> 
#> $emilia_3_8
#>     Token.surface
#>  1:            Ha
#>  2:             ,
#>  3:         meine
#>  4:        Mutter
#>  5:             !
#>  6:           Ich
#>  7:          hore
#>  8:         meine
#>  9:        Mutter
#> 10:             !
#> 
#> $der.prinz_4_1
#>      Token.surface
#>   1:        Kommen
#>   2:           Sie
#>   3:             ,
#>   4:     Marinelli
#>   5:             !
#>  ---              
#> 556:          Dass
#> 557:           Sie
#> 558:         Recht
#> 559:         haben
#> 560:             !
#> 
#> $marinelli_4_1
#>      Token.surface
#>   1:             O
#>   2:           der
#>   3:  mutterlichen
#>   4:           Wut
#>   5:             !
#>  ---              
#> 618:     verzeihen
#> 619:             ,
#> 620:      gnadiger
#> 621:          Herr
#> 622:             -
#> 
#> $battista_4_2
#>    Token.surface
#> 1:          Eben
#> 2:         kommt
#> 3:           die
#> 4:        Grafin
#> 5:            an
#> 6:             .
#> 7:        Orsina
#> 8:             .
#> 
#> $der.prinz_4_2
#>      Token.surface
#>   1:           Die
#>   2:        Grafin
#>   3:             ?
#>   4:           Was
#>   5:           fur
#>  ---              
#> 163:         Dinge
#> 164:          hier
#> 165:            zu
#> 166:           tun
#> 167:             -
#> 
#> $marinelli_4_2
#>      Token.surface
#>   1:           Ich
#>   2:      erstaune
#>   3:       daruber
#>   4:             ,
#>   5:         nicht
#>  ---              
#> 145:         ihrer
#> 146:        besten
#> 147:        Stunde
#> 148:   ausgefahren
#> 149:             .
#> 
#> $orsina_4_3
#>       Token.surface
#>    1:           Was
#>    2:           ist
#>    3:           das
#>    4:             ?
#>    5:             -
#>   ---              
#> 1125:           wir
#> 1126:        mussen
#> 1127:           uns
#> 1128:      sprechen
#> 1129:             -
#> 
#> $marinelli_4_3
#>      Token.surface
#>   1:           Der
#>   2:         Prinz
#>   3:             ,
#>   4:         meine
#>   5:       gnadige
#>  ---              
#> 172:             -
#> 173:          Aber
#> 174:       gnadige
#> 175:        Grafin
#> 176:             -
#> 
#> $der.prinz_4_4
#>     Token.surface
#>  1:           Ich
#>  2:          muss
#>  3:           ihm
#>  4:            zu
#>  5:         Hulfe
#>  6:        kommen
#>  7:             -
#>  8:          Sieh
#>  9:            da
#> 10:             !
#> 11:        unsere
#> 12:        schone
#> 13:        Grafin
#> 14:             .
#> 15:             -
#> 16:           Wie
#> 17:          sehr
#> 18:       betaure
#> 19:           ich
#> 20:             ,
#> 21:        Madame
#> 22:             ,
#> 23:          dass
#> 24:           ich
#> 25:           mir
#> 26:           die
#> 27:          Ehre
#> 28:         Ihres
#> 29:       Besuchs
#> 30:           fur
#> 31:         heute
#> 32:            so
#> 33:         wenig
#> 34:            zu
#> 35:         Nutze
#> 36:        machen
#> 37:          kann
#> 38:             !
#> 39:           Ich
#> 40:           bin
#> 41:  beschaftiget
#> 42:             .
#> 43:           Ich
#> 44:           bin
#> 45:         nicht
#> 46:        allein
#> 47:             .
#> 48:             -
#> 49:           Ein
#> 50:      andermal
#> 51:             ,
#> 52:         meine
#> 53:         liebe
#> 54:        Grafin
#> 55:             !
#> 56:           Ein
#> 57:      andermal
#> 58:             .
#> 59:             -
#> 60:          Itzt
#> 61:        halten
#> 62:           Sie
#> 63:        langer
#> 64:          sich
#> 65:         nicht
#> 66:           auf
#> 67:             .
#> 68:            Ja
#> 69:         nicht
#> 70:        langer
#> 71:             !
#> 72:             -
#> 73:           Und
#> 74:           Sie
#> 75:             ,
#> 76:     Marinelli
#> 77:             ,
#> 78:           ich
#> 79:       erwarte
#> 80:           Sie
#> 81:             .
#> 82:             -
#>     Token.surface
#> 
#> $orsina_4_4
#>    Token.surface
#> 1:            Ha
#> 2:             !
#> 3:            da
#> 4:           ist
#> 5:            er
#> 6:             .
#> 
#> $marinelli_4_5
#>      Token.surface
#>   1:         Haben
#>   2:           Sie
#>   3:            es
#>   4:             ,
#>   5:       gnadige
#>  ---              
#> 220:            um
#> 221:           den
#> 222:          Hals
#> 223:         reden
#> 224:             -
#> 
#> $orsina_4_5
#>      Token.surface
#>   1:           Hab
#>   2:             '
#>   3:           ich
#>   4:             ?
#>   5:           hab
#>  ---              
#> 934:             ,
#> 935:           der
#> 936:       eiligst
#> 937:   hereintritt
#> 938:             .
#> 
#> $odoardo_4_6
#>      Token.surface
#>   1:     Verzeihen
#>   2:           Sie
#>   3:             ,
#>   4:       gnadige
#>   5:          Frau
#>  ---              
#> 116:           nur
#> 117:             ,
#> 118:          mein
#> 119:          Herr
#> 120:             .
#> 
#> $orsina_4_6
#>      Token.surface
#>   1:           Ich
#>   2:          habe
#>   3:          hier
#>   4:        nichts
#>   5:            zu
#>  ---              
#> 115:             .
#> 116:           Ich
#> 117:       erwarte
#> 118:           ihn
#> 119:             .
#> 
#> $marinelli_4_6
#>      Token.surface
#>   1:           Nun
#>   2:      vollends
#>   3:             !
#>   4:           der
#>   5:          Alte
#>  ---              
#> 202:           ihr
#> 203:         nicht
#> 204:           ins
#> 205:          Wort
#> 206:             .
#> 
#> $orsina_4_7
#>      Token.surface
#>   1:           Was
#>   2:            er
#>   3:         Ihnen
#>   4:          auch
#>   5:            da
#>  ---              
#> 746:        werden
#> 747:             !
#> 748:           das
#> 749:        sollte
#> 750:             !
#> 
#> $odoardo_4_7
#>      Token.surface
#>   1: Unglucklicher
#>   2:             ?
#>   3:           Auf
#>   4:          mich
#>   5:        warten
#>  ---              
#> 397:           mit
#> 398:           mir
#> 399:            zu
#> 400:           tun
#> 401:             .
#> 
#> $claudia.galotti_4_8
#>      Token.surface
#>   1:       Erraten
#>   2:             !
#>   3:             -
#>   4:            Ah
#>   5:             ,
#>  ---              
#> 228:        ungern
#> 229:           von
#> 230:           dem
#> 231:         Kinde
#> 232:             .
#> 
#> $odoardo_4_8
#>      Token.surface
#>   1:           Gut
#>   2:             ,
#>   3:           gut
#>   4:             .
#>   5:           Sei
#>  ---              
#> 240:            Er
#> 241:         fuhrt
#> 242:           sie
#> 243:            ab
#> 244:             .
#> 
#> $orsina_4_8
#>     Token.surface
#>  1:           Nun
#>  2:             ,
#>  3:           hab
#>  4:             '
#>  5:           ich
#>  6:       gelogen
#>  7:             ?
#>  8:           Bin
#>  9:           ich
#> 10:    wahnwitzig
#> 11:             ?
#> 12:         Nicht
#> 13:        anders
#> 14:             .
#> 15:         Warum
#> 16:         nicht
#> 17:             ?
#> 18:          Sehr
#> 19:          gern
#> 20:             .
#> 
#> $marinelli_5_1
#>      Token.surface
#>   1:          Hier
#>   2:             ,
#>   3:      gnadiger
#>   4:          Herr
#>   5:             ,
#>  ---              
#> 325: Unschuldigste
#> 326:           von
#> 327:           der
#> 328:          Welt
#> 329:             !
#> 
#> $der.prinz_5_1
#>      Token.surface
#>   1:          Wenn
#>   2:            er
#>   3:           nun
#>   4:          aber
#>   5:            so
#>   6:          zahm
#>   7:         nicht
#>   8:           ist
#>   9:             ?
#>  10:           Und
#>  11:    schwerlich
#>  12:             ,
#>  13:    schwerlich
#>  14:          wird
#>  15:            er
#>  16:            es
#>  17:          sein
#>  18:             .
#>  19:           Ich
#>  20:         kenne
#>  21:           ihn
#>  22:            zu
#>  23:           gut
#>  24:             .
#>  25:             -
#>  26:          Wenn
#>  27:            er
#>  28:     hochstens
#>  29:        seinen
#>  30:       Argwohn
#>  31:      erstickt
#>  32:             ,
#>  33:         seine
#>  34:           Wut
#>  35:     verbeisst
#>  36:             :
#>  37:          aber
#>  38:       Emilien
#>  39:             ,
#>  40:       anstatt
#>  41:           sie
#>  42:          nach
#>  43:           der
#>  44:         Stadt
#>  45:            zu
#>  46:        fahren
#>  47:             ,
#>  48:           mit
#>  49:          sich
#>  50:         nimmt
#>  51:             ?
#>  52:           bei
#>  53:          sich
#>  54:        behalt
#>  55:             ?
#>  56:          oder
#>  57:          wohl
#>  58:           gar
#>  59:            in
#>  60:           ein
#>  61:       Kloster
#>  62:             ,
#>  63:        ausser
#>  64:        meinem
#>  65:       Gebiete
#>  66:             ,
#>  67:  verschliesst
#>  68:             ?
#>  69:           Wie
#>  70:          dann
#>  71:             ?
#>  72:          Wenn
#>  73:            er
#>  74:           nun
#>  75:          aber
#>  76:             !
#>  77:           Wie
#>  78:          dann
#>  79:             ?
#>  80:           Was
#>  81:          wird
#>  82:            es
#>  83:           uns
#>  84:          dann
#>  85:        helfen
#>  86:             ,
#>  87:          dass
#>  88:           der
#>  89:  ungluckliche
#>  90:          Graf
#>  91:          sein
#>  92:         Leben
#>  93:       daruber
#>  94:      verloren
#>  95:             ?
#>  96:           Nur
#>  97:             ,
#>  98:     Marinelli
#>  99:             !
#> 100:             -
#>      Token.surface
#> 
#> $odoardo_5_2
#>      Token.surface
#>   1:          Noch
#>   2:       niemand
#>   3:          hier
#>   4:             ?
#>   5:             -
#>  ---              
#> 208:         Holle
#> 209:             ,
#> 210:           und
#> 211:       erwache
#> 212:             !
#> 
#> $marinelli_5_3
#>      Token.surface
#>   1:            Wo
#>   2:       blieben
#>   3:           Sie
#>   4:             ,
#>   5:          mein
#>  ---              
#> 166:             '
#> 167:           und
#> 168:          hole
#> 169:           ihn
#> 170:             .
#> 
#> $odoardo_5_3
#>      Token.surface
#>   1:           War
#>   2:         meine
#>   3:       Tochter
#>   4:          hier
#>   5:             ?
#>  ---              
#> 149:           sie
#> 150:          muss
#> 151:           mit
#> 152:           mir
#> 153:             .
#> 
#> $odoardo_5_4
#>      Token.surface
#>   1:           Wie
#>   2:             ?
#>   3:             -
#>   4:    Nimmermehr
#>   5:             !
#>  ---              
#> 202:         alter
#> 203:         Knabe
#> 204:             ,
#> 205:         ruhig
#> 206:             !
#> 
#> $der.prinz_5_5
#>      Token.surface
#>   1:            Ah
#>   2:             ,
#>   3:          mein
#>   4:        lieber
#>   5:             ,
#>  ---              
#> 610:           und
#> 611:     Marinelli
#> 612:         gehen
#> 613:            ab
#> 614:             .
#> 
#> $odoardo_5_5
#>      Token.surface
#>   1:      Gnadiger
#>   2:          Herr
#>   3:             ,
#>   4:           ich
#>   5:         halte
#>  ---              
#> 661:           nur
#> 662:             ,
#> 663:      gnadiger
#> 664:          Herr
#> 665:             .
#> 
#> $marinelli_5_5
#>      Token.surface
#>   1:          Wenn
#>   2:           Sie
#>   3:          mich
#>   4:            so
#>   5:           gar
#>  ---              
#> 331:     besondere
#> 332:    Verwahrung
#> 333:            zu
#> 334:       bringen
#> 335:             .
#> 
#> $odoardo_5_6
#>      Token.surface
#>   1:         Warum
#>   2:         nicht
#>   3:             ?
#>   4:             -
#>   5:      Herzlich
#>  ---              
#> 195:             ;
#> 196:            er
#> 197:          will
#> 198:           sie
#> 199:             !
#> 
#> $emilia_5_7
#>      Token.surface
#>   1:           Wie
#>   2:             ?
#>   3:           Sie
#>   4:          hier
#>   5:             ,
#>  ---              
#> 724:             ,
#> 725:         diese
#> 726:    vaterliche
#> 727:          Hand
#> 728:             .
#> 
#> $odoardo_5_7
#>      Token.surface
#>   1:           Und
#>   2:            du
#>   3:            so
#>   4:         ruhig
#>   5:             ,
#>  ---              
#> 383:           sie
#> 384:            in
#> 385:         seine
#> 386:          Arme
#> 387:             .
#> 
#> $der.prinz_5_8
#>      Token.surface
#>   1:           Was
#>   2:           ist
#>   3:           das
#>   4:             ?
#>   5:             -
#>  ---              
#> 114:             ?
#> 115:          Ende
#> 116:           des
#> 117:  Trauerspiels
#> 118:             .
#> 
#> $odoardo_5_8
#>      Token.surface
#>   1:          Sehr
#>   2:          wohl
#>   3:             ;
#>   4:          sehr
#>   5:          wohl
#>  ---              
#> 185:           dem
#> 186:       Richter
#> 187:         unser
#> 188:         aller
#> 189:             !
#> 
#> $marinelli_5_8
#>    Token.surface
#> 1:           Weh
#> 2:           mir
#> 3:             !
#> 
#> $emilia_5_8
#>     Token.surface
#>  1:         Nicht
#>  2:           Sie
#>  3:             ,
#>  4:          mein
#>  5:         Vater
#>  6:             -
#>  7:           Ich
#>  8:        selbst
#>  9:             -
#> 10:           ich
#> 11:        selbst
#> 12:             -
#> 13:            Ah
#> 14:             -
#> 15:          mein
#> 16:         Vater
#> 17:             -
#> 18:           Sie
#> 19:        stirbt
#> 20:             ,
#> 21:           und
#> 22:            er
#> 23:          legt
#> 24:           sie
#> 25:         sanft
#> 26:           auf
#> 27:           den
#> 28:         Boden
#> 29:             .
#>     Token.surface
#>