Load drama

This function loads one or more of the installed plays and returns them as a QDDrama object.

The function combine(x, y) can be used to merge multiple objects of the type QDDrama into one.

The function split(x) expects an object of type QDDrama and can be used to split a QDDrama object that consists of multiple dramas into a list thereof. It is the counterpart to combine(x, y).

The function numberOfPlays() determines how many different plays are contained in a single QDDrama object.

loadDrama(ids, defaultCollection = "qd")

combine(x, y)

# S3 method for QDDrama
split(x, ...)

numberOfPlays(x)

Arguments

ids
A vector of ids.

defaultCollection
If the ids do not have a collection prefix, the defaultCollection prefix is applied.

x
The object of class QDDrama (consisting of multiple dramas). For split() it should consistof multiple plays. For combine() it can but doesn't have to.

y
A QDDrama

...
All other arguments are ignored.

Value

The function returns a QDDrama object. This is essentially a list of data.tables, covering the different aspects (utterances, segments, characters, ...). If multiple ids have been supplied as arguments, the tables contain the information of multiple plays.

Examples

# both are equivalent

# NOT RUN {
d <- loadDrama(c("test:rksp.0", "test:rjmw.0"))
d <- loadDrama(c("rksp.0", "rjmw.0"), defaultCollection = "test")
# }

data(rksp.0)
data(rjmw.0)
d <- combine(rjmw.0, rksp.0)
data(rksp.0)
data(rjmw.0)
d <- combine(rjmw.0, rksp.0)
dlist <- split(d)
# returns 1
numberOfPlays(rksp.0)

#> [1] 1

# returns 2
numberOfPlays(combine(rksp.0, rjmw.0))

#> [1] 2